requirements.txt
aiohttp>=3.9.1
pillow>=10.0.0
pandas>=2.1.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
flask>=3.0.0
werkzeug>=3.0.0
openpyxl>=3.1.0
